+++ lib/features/order_list/presentation/widgets/new_order.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:picstate/custom_widgets/rounded_button.dart';
-import 'package:picstate/custom_widgets/text_input.dart';
+import 'package:picstate/core/widgets/rounded_button.dart';
+import 'package:picstate/core/widgets/text_input.dart';
 import 'package:picstate/config/constants.dart';
 import 'package:picstate/logic/logic.dart';
