+++ lib/features/todo_list/presentation/widgets/new_task.dart
@@ -2,8 +2,8 @@ import 'package:date_picker_timeline/date_picker_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:picstate/config/constants.dart';
-import 'package:picstate/custom_widgets/rounded_button.dart';
-import 'package:picstate/custom_widgets/text_input.dart';
+import 'package:picstate/core/widgets/rounded_button.dart';
+import 'package:picstate/core/widgets/text_input.dart';
 import 'package:picstate/logic/logic.dart';
 
 class NewTask extends StatefulWidget {
